Added hidden optional positional dummy argument to consume redundant `modules` argument when executing via `cargo modules`
